requirements.txt
numpy
PyQt5
pyqtgraph
pyserial
PyYAML
scipy
six
pandas